bug fix: apply the same damage conversion on the xwayland path
xwayland-xdg-shell converted surface damage with the same bug wprsd had:
buffer_scale only, and to_buffer's transform area given the damage rect's
own size instead of the surface size.

Move surface_damage_to_buffer into compositor_utils and use it from both
paths. Xwayland surfaces have no viewport, so the viewport argument is
None there and this only corrects the transform area, but it keeps the
two conversions from drifting apart again.
