+++ src/server/smithay_handlers.rs
@@ -58,14 +58,9 @@ use smithay::reexports::wayland_server::protocol::wl_surface::WlSurface;
 use smithay::reexports::wayland_server::Client;
 use smithay::reexports::wayland_server::Resource;
 use smithay::reexports::wayland_server::WEnum;
-use smithay::utils::Buffer as BufferCoords;
 use smithay::utils::Logical;
 use smithay::utils::Point;
-use smithay::utils::Point as SmithayPoint;
-use smithay::utils::Rectangle as SmithayRectangle;
 use smithay::utils::Serial;
-use smithay::utils::Size as SmithaySize;
-use smithay::utils::Transform as SmithayTransform;
 use smithay::wayland::buffer::BufferHandler;
 use smithay::wayland::compositor;
 use smithay::wayland::compositor::BufferAssignment as SmithayBufferAssignment;
@@ -123,7 +118,6 @@ use crate::serialization::wayland::SurfaceRequest;
 use crate::serialization::wayland::SurfaceRequestPayload;
 use crate::serialization::wayland::SurfaceState;
 use crate::serialization::wayland::Transform;
-use crate::serialization::wayland::ViewportState;
 use crate::serialization::wayland::WlSurfaceId;
 use crate::serialization::xdg_shell::DecorationMode;
 use crate::serialization::xdg_shell::Move;
@@ -686,67 +680,6 @@ pub fn set_viewport_state(viewport_state: &ViewportCachedState, surface_state: &
     surface_state.viewport_state = Some(viewport_state.into());
 }
 
-/// Convert a damage rectangle from surface coordinates to buffer coordinates.
-///
-/// The viewport is part of this mapping, not just buffer_scale: chromium leaves
-/// buffer_scale at 1 and scales entirely through the viewport, so using
-/// buffer_scale alone makes this the identity and the damage then covers only
-/// the top-left corner of the buffer.
-fn surface_damage_to_buffer(
-    rect: SmithayRectangle<i32, Logical>,
-    buffer_scale: i32,
-    transform: SmithayTransform,
-    viewport_state: Option<&ViewportState>,
-    buffer_size: Option<(i32, i32)>,
-) -> SmithayRectangle<i32, BufferCoords> {
-    let src = viewport_state.and_then(|viewport_state| viewport_state.src);
-    let dst = viewport_state.and_then(|viewport_state| viewport_state.dst);
-
-    // Surface-local size the buffer covers before the destination scaling.
-    let surface_size = match (src, buffer_size) {
-        (Some(src), _) => (src.size.w, src.size.h),
-        (None, Some((width, height))) => {
-            // A rotating transform swaps the axes between the two spaces.
-            let (width, height) = match transform {
-                SmithayTransform::_90
-                | SmithayTransform::_270
-                | SmithayTransform::Flipped90
-                | SmithayTransform::Flipped270 => (height, width),
-                _ => (width, height),
-            };
-            let scale = f64::from(buffer_scale.max(1));
-            (f64::from(width) / scale, f64::from(height) / scale)
-        },
-        // Nothing to scale against.
-        (None, None) => return rect.to_buffer(buffer_scale, transform, &rect.size),
-    };
-
-    let (scale_w, scale_h) = match dst {
-        Some(dst) if dst.w > 0 && dst.h > 0 => (
-            surface_size.0 / f64::from(dst.w),
-            surface_size.1 / f64::from(dst.h),
-        ),
-        _ => (1.0, 1.0),
-    };
-    let (offset_x, offset_y) = src.map_or((0.0, 0.0), |src| (src.loc.x, src.loc.y));
-
-    let rect = rect.to_f64();
-    let unviewported = SmithayRectangle::new(
-        SmithayPoint::from((
-            rect.loc.x * scale_w + offset_x,
-            rect.loc.y * scale_h + offset_y,
-        )),
-        SmithaySize::from((rect.size.w * scale_w, rect.size.h * scale_h)),
-    );
-
-    // Round outwards: too much damage costs a repaint, too little leaves stale
-    // pixels.
-    let area = SmithaySize::from((surface_size.0.ceil() as i32, surface_size.1.ceil() as i32));
-    unviewported
-        .to_i32_up::<i32>()
-        .to_buffer(buffer_scale, transform, &area)
-}
-
 #[instrument(skip_all, level = "debug")]
 pub fn set_xdg_surface_attributes(surface_data: &SurfaceData, surface_state: &mut SurfaceState) {
     if surface_data.cached_state.has::<SurfaceCachedState>() {
@@ -955,7 +888,7 @@ pub fn commit_impl(
         .iter()
         .map(|damage| match damage {
             Damage::Buffer(rect) => *rect,
-            Damage::Surface(rect) => surface_damage_to_buffer(
+            Damage::Surface(rect) => compositor_utils::surface_damage_to_buffer(
                 *rect,
                 surface_state.buffer_scale,
                 surface_state
